refactor: #138 RootPage를 Layout으로 이름 변경
